@@ -5,7 +5,6 @@ require (
 	github.com/docker/go-connections v0.5.0
 	github.com/johntdyer/slackrus v0.0.0-20230315191314-80bc92dee4fc
 	github.com/opencontainers/image-spec v1.1.1
-	github.com/pkg/errors v0.9.1
 	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.10.0
 	github.com/urfave/cli v1.22.16
@@ -26,6 +25,7 @@ require (
 	github.com/morikuni/aec v1.0.0 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/opencontainers/go-digest v1.0.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/stretchr/objx v0.5.2 // indirect
@@ -38,4 +38,4 @@ require (
 	gotest.tools/v3 v3.0.2 // indirect
 )
 
-go 1.24
+go 1.26
