@@ -6,7 +6,6 @@ import (
 
 	"github.com/alexei-led/pumba/pkg/chaos"
 	"github.com/alexei-led/pumba/pkg/chaos/netem"
-	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 )
 
@@ -45,12 +44,12 @@ func (cmd *lossContext) loss(c *cli.Context) error {
 	// parse common chaos flags
 	globalParams, err := chaos.ParseGlobalParams(c)
 	if err != nil {
-		return errors.Wrap(err, "error parsing global parameters")
+		return fmt.Errorf("error parsing global parameters: %w", err)
 	}
 	// parse netem flags
 	netemParams, err := parseNetemParams(c.Parent(), globalParams.Interval)
 	if err != nil {
-		return errors.Wrap(err, "error parsing netem parameters")
+		return fmt.Errorf("error parsing netem parameters: %w", err)
 	}
 	// get loss percentage
 	percent := c.Float64("percent")
@@ -59,12 +58,12 @@ func (cmd *lossContext) loss(c *cli.Context) error {
 	// init netem loss command
 	lossCommand, err := netem.NewLossCommand(chaos.DockerClient, globalParams, netemParams, percent, correlation)
 	if err != nil {
-		return errors.Wrap(err, "error creating netem loss command")
+		return fmt.Errorf("error creating netem loss command: %w", err)
 	}
 	// run netem command
 	err = chaos.RunChaosCommand(cmd.context, lossCommand, globalParams)
 	if err != nil {
-		return errors.Wrap(err, "error running netem loss command")
+		return fmt.Errorf("error running netem loss command: %w", err)
 	}
 	return nil
 }
