@@ -2,13 +2,14 @@ package netem
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"regexp"
 	"strconv"
 	"sync"
 
 	"github.com/alexei-led/pumba/pkg/chaos"
 	"github.com/alexei-led/pumba/pkg/container"
-	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -17,7 +18,7 @@ func parseRate(rate string) (string, error) {
 	reRate := regexp.MustCompile(`\d+[gmk]?bit`)
 	validRate := reRate.FindString(rate)
 	if rate != validRate {
-		return "", errors.Errorf("invalid rate, must match '%s'", reRate.String())
+		return "", fmt.Errorf("invalid rate, must match '%s'", reRate.String())
 	}
 	return rate, nil
 }
@@ -46,7 +47,7 @@ func NewRateCommand(client container.Client,
 	}
 	rate, err := parseRate(rate)
 	if err != nil {
-		return nil, errors.Wrap(err, "invalid rate")
+		return nil, fmt.Errorf("invalid rate: %w", err)
 	}
 
 	// validate cell size
@@ -75,7 +76,7 @@ func (n *rateCommand) Run(ctx context.Context, random bool) error {
 	}).Debug("listing matching containers")
 	containers, err := container.ListNContainers(ctx, n.client, n.names, n.pattern, n.labels, n.limit)
 	if err != nil {
-		return errors.Wrap(err, "error listing containers")
+		return fmt.Errorf("error listing containers: %w", err)
 	}
 	if len(containers) == 0 {
 		log.Warning("no containers found")
@@ -136,7 +137,7 @@ func (n *rateCommand) Run(ctx context.Context, random bool) error {
 	for _, err = range errs {
 		// take first found error
 		if err != nil {
-			return errors.Wrap(err, "failed to set network rate for one or more containers")
+			return fmt.Errorf("failed to set network rate for one or more containers: %w", err)
 		}
 	}
 
