@@ -9,7 +9,7 @@ import (
 // CreateTestContainers create test container
 func CreateTestContainers(count int) []*Container {
 	var containers []*Container
-	for i := 0; i < count; i++ {
+	for i := range count {
 		containers = append(containers, &Container{
 			DetailsResponse(AsMap("Name", fmt.Sprintf("c%d", i))),
 			ImageDetailsResponse(AsMap()),
@@ -21,7 +21,7 @@ func CreateTestContainers(count int) []*Container {
 // CreateLabeledTestContainers generate test containers with labels
 func CreateLabeledTestContainers(count int, labels map[string]string) []*Container {
 	var containers []*Container
-	for i := 0; i < count; i++ {
+	for i := range count {
 		containers = append(containers, &Container{
 			DetailsResponse(AsMap("Name", fmt.Sprintf("c%d", i), "Labels", labels)),
 			ImageDetailsResponse(AsMap()),
